requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
line-bot-sdk==1.19.0
MarkupSafe==1.1.1
pytz==2021.1
requests==2.25.1
urllib3==1.26.4
uWSGI==2.0.19.1
Werkzeug==1.0.1
